refactored module naming
