+++ py_image_dedup/library/deduplication_result.py
