+++ py_image_dedup/persistence/elasticsearchstorebackend.py
