+++ py_image_dedup/persistence/metadata_key.py
