+++ tests/test_select_images_to_delete.py
@@ -2,9 +2,9 @@ import unittest
 from random import shuffle
 from random import uniform
 
-from py_image_dedup.library.DeduplicatorConfig import DeduplicatorConfig
-from py_image_dedup.library.ImageMatchDeduplicator import ImageMatchDeduplicator
-from py_image_dedup.persistence.MetadataKey import MetadataKey
+from py_image_dedup.library.deduplicator_config import DeduplicatorConfig
+from py_image_dedup.library.image_match_deduplicator import ImageMatchDeduplicator
+from py_image_dedup.persistence.metadata_key import MetadataKey
 from tests import TestBase
